add: endpints

+++ src/main/java/com/openlib/controller/AdminController.java
@@ -39,7 +39,7 @@ public class AdminController {
         return null;
     }
 
-    public boolean deleteBook(int bookId) {
+    public boolean deleteBook(Long bookId) {
         return store.deleteBook(bookId);
     }
